requirements.txt
Django==1.10.4
django-annoying==0.10.3
psycopg2==2.6.2
requests==2.12.4
beautifulsoup4==4.5.1
